+++ tests/test_synchronization.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import asyncio
+from typing import Any
 
 import pytest
 
@@ -12,6 +13,8 @@ from anyio import (
     Semaphore,
     WouldBlock,
     create_task_group,
+    fail_after,
+    run,
     to_thread,
     wait_all_tasks_blocked,
 )
@@ -140,6 +143,21 @@ class TestLock:
         task1.cancel()
         await asyncio.wait_for(task2, 1)
 
+    def test_instantiate_outside_event_loop(
+        self, anyio_backend_name: str, anyio_backend_options: dict[str, Any]
+    ) -> None:
+        async def use_lock() -> None:
+            async with lock:
+                pass
+
+        lock = Lock()
+        statistics = lock.statistics()
+        assert not statistics.locked
+        assert statistics.owner is None
+        assert statistics.tasks_waiting == 0
+
+        run(use_lock, backend=anyio_backend_name, backend_options=anyio_backend_options)
+
 
 class TestEvent:
     async def test_event(self) -> None:
@@ -206,6 +224,21 @@ class TestEvent:
             event.set()
 
         assert event.statistics().tasks_waiting == 0
+
+    def test_instantiate_outside_event_loop(
+        self, anyio_backend_name: str, anyio_backend_options: dict[str, Any]
+    ) -> None:
+        async def use_event() -> None:
+            event.set()
+            await event.wait()
+
+        event = Event()
+        assert not event.is_set()
+        assert event.statistics().tasks_waiting == 0
+
+        run(
+            use_event, backend=anyio_backend_name, backend_options=anyio_backend_options
+        )
 
 
 class TestCondition:
@@ -302,6 +335,22 @@ class TestCondition:
 
         assert not condition.statistics().lock_statistics.locked
         assert condition.statistics().tasks_waiting == 0
+
+    def test_instantiate_outside_event_loop(
+        self, anyio_backend_name: str, anyio_backend_options: dict[str, Any]
+    ) -> None:
+        async def use_condition() -> None:
+            async with condition:
+                pass
+
+        condition = Condition()
+        assert condition.statistics().tasks_waiting == 0
+
+        run(
+            use_condition,
+            backend=anyio_backend_name,
+            backend_options=anyio_backend_options,
+        )
 
 
 class TestSemaphore:
@@ -424,6 +473,22 @@ class TestSemaphore:
         await asyncio.sleep(0)
         task1.cancel()
         await asyncio.wait_for(task2, 1)
+
+    def test_instantiate_outside_event_loop(
+        self, anyio_backend_name: str, anyio_backend_options: dict[str, Any]
+    ) -> None:
+        async def use_semaphore() -> None:
+            async with semaphore:
+                pass
+
+        semaphore = Semaphore(1)
+        assert semaphore.statistics().tasks_waiting == 0
+
+        run(
+            use_semaphore,
+            backend=anyio_backend_name,
+            backend_options=anyio_backend_options,
+        )
 
 
 class TestCapacityLimiter:
@@ -564,3 +629,63 @@ class TestCapacityLimiter:
             event.set()
 
         assert results == [0, 1, 2]
+
+    async def test_increase_tokens_lets_others_acquire(self) -> None:
+        limiter = CapacityLimiter(1)
+        entered_events = [Event() for _ in range(3)]
+        continue_event = Event()
+
+        async def worker(entered_event: Event) -> None:
+            async with limiter:
+                entered_event.set()
+                await continue_event.wait()
+
+        async with create_task_group() as tg:
+            for event in entered_events[:2]:
+                tg.start_soon(worker, event)
+
+            # One task should be able to acquire the limiter while the other is left
+            # waiting
+            await wait_all_tasks_blocked()
+            assert sum(ev.is_set() for ev in entered_events) == 1
+
+            # Increase the total tokens and start another worker.
+            # All tasks should be able to acquire the limiter now.
+            limiter.total_tokens = 3
+            tg.start_soon(worker, entered_events[2])
+            with fail_after(1):
+                for ev in entered_events[1:]:
+                    await ev.wait()
+
+            # Allow all tasks to exit
+            continue_event.set()
+
+    def test_instantiate_outside_event_loop(
+        self, anyio_backend_name: str, anyio_backend_options: dict[str, Any]
+    ) -> None:
+        async def use_limiter() -> None:
+            async with limiter:
+                pass
+
+        limiter = CapacityLimiter(1)
+        limiter.total_tokens = 2
+
+        with pytest.raises(TypeError):
+            limiter.total_tokens = "2"  # type: ignore[assignment]
+
+        with pytest.raises(TypeError):
+            limiter.total_tokens = 3.0
+
+        assert limiter.total_tokens == 2
+        assert limiter.borrowed_tokens == 0
+        statistics = limiter.statistics()
+        assert statistics.total_tokens == 2
+        assert statistics.borrowed_tokens == 0
+        assert statistics.borrowers == ()
+        assert statistics.tasks_waiting == 0
+
+        run(
+            use_limiter,
+            backend=anyio_backend_name,
+            backend_options=anyio_backend_options,
+        )